Now I use Morphia, so the MongoDB driver is no longer needed.
Created classes Trip and Segment. The trip has name, definition if public and a list of segments. A segment consists of information for the start point, end point and transport.

The client is able to mark a trip with several segments and save them do the database as a trip with an array of segments.